refactor tests, clean up code
