+++ examples/OptionsPatternMvc.Example/Startup.cs
@@ -5,7 +5,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using OptionsPatternMvc.Example.Settings;
 using OptionsPatternValidation;
-using OptionsPatternValidation.Extensions;
 
 namespace OptionsPatternMvc.Example
 {
