+++ test/OptionsPatternValidation.Tests/Json/Unvalidated/SimpleSettings1.json
