+++ test/OptionsPatternValidation.Tests/Json/Unvalidated/SimpleSettings2.json
