+++ test/OptionsPatternValidation.Tests/OptionsPatternValidation.Tests.csproj
@@ -26,10 +26,10 @@
 
   <ItemGroup>
     <EmbeddedResource Include="TestSettingsJson\*.json" />
-    <EmbeddedResource Include="TestSettingsJson\AttributeValidated\Test1.json" />
-    <EmbeddedResource Include="TestSettingsJson\AttributeValidated\Test2.json" />
-    <EmbeddedResource Include="TestSettingsJson\Unvalidated\SimpleSettings1.json" />
-    <EmbeddedResource Include="TestSettingsJson\Unvalidated\SimpleSettings2.json" />
+    <EmbeddedResource Include="Json\AttributeValidated\Test1.json" />
+    <EmbeddedResource Include="Json\AttributeValidated\Test2.json" />
+    <EmbeddedResource Include="Json\Unvalidated\SimpleSettings1.json" />
+    <EmbeddedResource Include="Json\Unvalidated\SimpleSettings2.json" />
   </ItemGroup>
 
 </Project>
