+++ test/OptionsPatternValidation.Tests/ServiceCollectionExtensions/AddEagerlyValidatedSettingsTests.cs
@@ -1,0 +1,7 @@
+namespace OptionsPatternValidation.Tests.ServiceCollectionExtensions
+{
+    public class AddEagerlyValidatedSettingsTests
+    {
+        
+    }
+}
