+++ test/OptionsPatternValidation.Tests/ServiceCollectionExtensions/AddValidatedSettingsTests.cs
@@ -1,9 +1,9 @@
 using System;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using OptionsPatternValidation.Tests.TestHelpers;
-using OptionsPatternValidation.Tests.TestSettings.AttributeValidation;
-using OptionsPatternValidation.Tests.TestSettingsJson;
+using OptionsPatternValidation.Tests.Helpers;
+using OptionsPatternValidation.Tests.Json;
+using OptionsPatternValidation.Tests.Settings.AttributeValidation;
 using Xunit;
 
 namespace OptionsPatternValidation.Tests.ServiceCollectionExtensions
